refactor & turn into package & memory FS option
