+++ constants/HEADER_FORMAT_STRING.py
@@ -1,0 +1,1 @@
+HEADER_FORMAT_STRING = '<7i4h8i2h4i13h2H31h6f'
