+++ constants/TRACE_HEADER_SIZE.py
@@ -1,0 +1,1 @@
+TRACE_HEADER_SIZE = 240  # in bytes
