+++ testImport.py
@@ -1,0 +1,3 @@
+# Import classes from your brand new package
+from services.readsu import readsu as r
+r("paaa")
